requirements.txt
pyautogui==0.9.54
pyperclip==1.8.2
playwright>=1.40.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
pillow>=10.1.0
markdownify>=1.2.0
requests>=2.31.0
